ReviewService validates productLink using both methods
The ReviewService (which is hosted and versioned completely
independently from ProductService) has business logic that enforces that
a `productLink` must exist for a review to be considered valid.

Our business logic (inside ReviewService) showcases two possible ways to
do this remote API call: via a node-selector forwarding query or via a
DNS query + direct GET on the host that supports ProductService.

I added logic to ensure that BOTH methods confirm that the product
exists before allowing a `Review` `POST` to be considered "complete"