Automatically Join as OBSERVER
In order for the `reviewHost-8001` to be able to query "products" using
the "product" node group selector, it must join the `productHost-8000`'s
`/core/node-groups/product` as an OBSERVER. Before this commit, this was
a manual step executed using the REST API.

Now, when `ReviewHost` starts, it will automatically kick off a new
JoinProductNodeGroupTaskService ... which dynamically finds a host that
is running `/products` and issues the JoinPeerRequest for you. No more
manual steps!